# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LecturePage {
  pageNumber: number;
  text: string;
}

export type LectureSourceType = typeof LectureSourceType[keyof typeof LectureSourceType];


export const LectureSourceType = {
  pdf: 'pdf',
  image: 'image',
} as const;

export interface LectureSource {
  id: string;
  name: string;
  type?: LectureSourceType;
  base64Data?: string;
  mimeType?: string;
  text: string;
  pageCount: number;
  pages: LecturePage[];
}

export interface PageReference {
  sourceId: string;
  sourceName: string;
  pageNumber: number;
  excerpt: string;
}

export interface ExplainableTerm {
  term: string;
  contextSnippet: string;
  plainDefinition: string;
}

export type StudyModel = typeof StudyModel[keyof typeof StudyModel];


export const StudyModel = {
  'gemini-31-flash-lite': 'gemini-3.1-flash-lite',
  'gemini-37-flash': 'gemini-3.7-flash',
  'gemini-31-pro-preview': 'gemini-3.1-pro-preview',
  'gpt-56-luna': 'gpt-5.6-luna',
  'gpt-56-terra': 'gpt-5.6-terra',
  'gpt-56-sol': 'gpt-5.6-sol',
  'claude-haiku-4-5': 'claude-haiku-4-5',
  'claude-sonnet-5': 'claude-sonnet-5',
  'claude-opus-5': 'claude-opus-5',
  'claude-fable-5-1': 'claude-fable-5-1',
} as const;

export interface StudyExplanationInput {
  prompt: string;
  model?: StudyModel;
  sources: LectureSource[];
}

export interface StudyExplanation {
  title: string;
  answerMarkdown: string;
  terms: ExplainableTerm[];
  sourceRefs: PageReference[];
  generatedAt: string;
}

export interface TechnicalConceptInput {
  term: string;
  context: string;
  model?: StudyModel;
  sources: LectureSource[];
}

export interface TechnicalConceptExplanation {
  title: string;
  answerMarkdown: string;
  prerequisiteTerms: string[];
  sourceRefs: PageReference[];
}

export interface SelectedPassageInput {
  /**
     * @minLength 1
     * @maxLength 3000
     */
  selectedText: string;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  question: string;
  /** @maxLength 20000 */
  answerContext: string;
  model?: StudyModel;
  sources: LectureSource[];
}

export interface SelectedPassageExplanation {
  title: string;
  answerMarkdown: string;
  terms: ExplainableTerm[];
  sourceRefs: PageReference[];
}