# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: study
    description: AI explanations grounded in lecture sources
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /study/explain:
    post:
      operationId: explainStudyTopic
      tags: [study]
      summary: Explain a topic and identify technical concepts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyExplanationInput"
      responses:
        "200":
          description: Topic explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyExplanation"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Explanation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /study/concepts/explain:
    post:
      operationId: explainTechnicalConcept
      tags: [study]
      summary: Explain one technical concept in depth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechnicalConceptInput"
      responses:
        "200":
          description: Concept explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechnicalConceptExplanation"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Explanation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /study/follow-ups/explain:
    post:
      operationId: explainSelectedPassage
      tags: [study]
      summary: Answer a question about a selected passage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectedPassageInput"
      responses:
        "200":
          description: Contextual follow-up explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SelectedPassageExplanation"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Explanation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    LecturePage:
      type: object
      properties:
        pageNumber:
          type: integer
        text:
          type: string
      required: [pageNumber, text]
    LectureSource:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [pdf, image]
        base64Data:
          type: string
        mimeType:
          type: string
        text:
          type: string
        pageCount:
          type: integer
        pages:
          type: array
          items:
            $ref: "#/components/schemas/LecturePage"
      required: [id, name, text, pageCount, pages]
    PageReference:
      type: object
      properties:
        sourceId:
          type: string
        sourceName:
          type: string
        pageNumber:
          type: integer
        excerpt:
          type: string
      required: [sourceId, sourceName, pageNumber, excerpt]
    ExplainableTerm:
      type: object
      properties:
        term:
          type: string
        contextSnippet:
          type: string
        plainDefinition:
          type: string
      required: [term, contextSnippet, plainDefinition]
    StudyModel:
      type: string
      enum:
        - gemini-3.1-flash-lite
        - gemini-3.7-flash
        - gemini-3.1-pro-preview
        - gpt-5.6-luna
        - gpt-5.6-terra
        - gpt-5.6-sol
        - claude-haiku-4-5
        - claude-sonnet-5
        - claude-opus-5
        - claude-fable-5-1
    StudyExplanationInput:
      type: object
      properties:
        prompt:
          type: string
        model:
          $ref: "#/components/schemas/StudyModel"
        sources:
          type: array
          items:
            $ref: "#/components/schemas/LectureSource"
      required: [prompt, sources]
    StudyExplanation:
      type: object
      properties:
        title:
          type: string
        answerMarkdown:
          type: string
        terms:
          type: array
          items:
            $ref: "#/components/schemas/ExplainableTerm"
        sourceRefs:
          type: array
          items:
            $ref: "#/components/schemas/PageReference"
        generatedAt:
          type: string
      required: [title, answerMarkdown, terms, sourceRefs, generatedAt]
    TechnicalConceptInput:
      type: object
      properties:
        term:
          type: string
        context:
          type: string
        model:
          $ref: "#/components/schemas/StudyModel"
        sources:
          type: array
          items:
            $ref: "#/components/schemas/LectureSource"
      required: [term, context, sources]
    TechnicalConceptExplanation:
      type: object
      properties:
        title:
          type: string
        answerMarkdown:
          type: string
        prerequisiteTerms:
          type: array
          items:
            type: string
        sourceRefs:
          type: array
          items:
            $ref: "#/components/schemas/PageReference"
      required: [title, answerMarkdown, prerequisiteTerms, sourceRefs]
    SelectedPassageInput:
      type: object
      properties:
        selectedText:
          type: string
          minLength: 1
          maxLength: 3000
        question:
          type: string
          minLength: 1
          maxLength: 1000
        answerContext:
          type: string
          maxLength: 20000
        model:
          $ref: "#/components/schemas/StudyModel"
        sources:
          type: array
          items:
            $ref: "#/components/schemas/LectureSource"
      required: [selectedText, question, answerContext, sources]
    SelectedPassageExplanation:
      type: object
      properties:
        title:
          type: string
        answerMarkdown:
          type: string
        terms:
          type: array
          items:
            $ref: "#/components/schemas/ExplainableTerm"
        sourceRefs:
          type: array
          items:
            $ref: "#/components/schemas/PageReference"
      required: [title, answerMarkdown, terms, sourceRefs]